Genomic Prediction with cross validation

The bwgs.cv function carries out cross-validation using genotypic and phenotypic data from a reference population, with options for genotypic matrix processing and genomic breeding value estimation.

bwgs.cv(
  geno,
  pheno,
  FIXED = "NULL",
  MAXNA = 0.2,
  MAF = 0.05,
  pop.reduct.method = "NULL",
  sample.pop.size = "NULL",
  geno.reduct.method = "NULL",
  reduct.marker.size = "NULL",
  pval = "NULL",
  r2 = "NULL",
  MAP = "NULL",
  geno.impute.method = "NULL",
  predict.method = "NULL",
  nFolds,
  nTimes
)

Arguments

geno
Matrix (n x m) of genotypes for the training population: n lines with m markers. Genotypes should be coded -1, 0, 1. Missing data are allowed and coded as NA.

pheno
Vector (n x 1) of "phenotypes", i.e. observations or pre-processed, corrected values. This vector should have no missing values, otherwise missing values (NA) will be omitted in both pheno and geno. In a first step, bwgs.cv checks whether rownames(geno) match with names(pheno). If not the case, the common elements (intersect) are selected in both geno and pheno for further analyses. If a MAP file is provided, the selected set of markers are also sorted out in MAP.

FIXED
A matrix of fixed effect, to be used with some methods such as those included in BGLR, MUST have same rownames as geno and coded(-1 0 1)

MAXNA
The maximum proportion of missing value which is admitted for filtering marker columns in geno. Default value is 0.2

MAF
The minimum allele frequency for filtering marker colums in geno;default value is 0.05

pop.reduct.method
Method for reducing the size of the training population. Can be used for teaching purposes, no real interest in real life if the entire population is already genotyped and phenotyped. Default value is NULL (all training set used). Proposed methods are:

RANDOM: a subset of sample.pop.size is randomly selected for training the model, and the unselected part of the population is used for validation. The process is repeated nFolds * nTimes to have the same number of replicates than with cross-validation.

OPTI: the optimization algorithm based on CDmean (Rincent et al 2012) to select a subset which maximizes average CD (coefficient of determination) in the validation set. Since the process is long and has some stochastic component, it is repeated only nTimes.

sample.pop.size
The size of the subset of individuals in the training set (both geno and pheno) selected by pop.reduct.method if not NULL.

geno.reduct.method
Allows sampling a subset of markers for speeding up computing time and/or avoid introducing more noise than informative markers. Options are:

RMR: Random sampling (without replacement) of a subset of markers. To be used with the parameter “reduct.marker.size”.

LD (with r2 and MAP): enables “pruning” of markers which are in LD > r2. Only the marker with the least missing values is kept for each pair in LD>r2. To allow faster computation, r2 is estimated chromosome by chromosome, so a MAP file is required with information of marker assignation to chromosomes. The MAP file should contain at least three columns: marker_name, chromosome_name and distance_from_origin (either genetic of physical distance, only used for sorting markers, LD being re-estimated from marker Data).

ANO (with pval): one-way ANOVA are carried out with R function lm on trait “pheno”. Every markers are tested one at a time, and only markers with pvalue<pval are kept for GEBV prediction.

ANO+LD (with pval and r2, MAP is facultative): combines a first step of marker selection with ANO, then a second step of pruning using LD option.

reduct.marker.size
Specifies the number of markers for the genotypic reduction using RMR (reduct.size < m).

pval
p value for ANO method, 0 < pval < 1.

r2
Coefficient of linkage disequilibrium (LD). Setting 0<r2<1 if the genotypic reduction method is in LD or ANO+LD .

MAP
A matrix with markers in rows and at least ONE columns with colnames= "chrom". Used for computing r2 within linkage groups.

geno.impute.method
Allow missing marker data imputation using the two methods proposed in function A.mat of package rrBLUP, namely:

MNI: missing data are replaced by the mean allele frequency of the marker (column in geno)

EMI: missing data are replaced using an expectation-maximization methods described in function A.mat (Endelman & Janninck 2012).

Default value is NULL.

Note that these imputation methods are only suited when there are a few missing value, typically in marker data from SNP chips of KasPAR. They are NOT suited for imputing marker data from low density to high density designs, and when there are MANY missing Data as typically provided by GBS. More sophisticated software (e.g. Beagles, Browning & Browning 2016) should be used before BWGS.

predict.method
The options for genomic breeding value prediction methods. The available options are:

GBLUP: performs G-BLUP using a marker-based relationship matrix, implemented through BGLR R-library. Equivalent to ridge regression (RR-BLUP) of marker effects.

EGBLUP: performs EG-BLUP, i.e. BLUP using a "squared" relationship matrix to model epistatic 2x2 interactions, as described by Jiang & Reif (2015), using BGLR library

RR: ridge regression, using package glmnet. In theory, strictly equivalent to gblup.

LASSO: Least Absolute Shrinkage and Selection Operator is another penalized regression methods which yield more shrinked estimates than RR. Run by glmnet library.

EN: Elastic Net (Zou and Hastie, 2005), which is a weighted combination of RR and LASSO, using glmnet library

Several Bayesian methods, using the BGLR library:

BRR: Bayesian ridge regression: same as rr-blup, but bayesian resolution. Induces homogeneous shrinkage of all markers effects towards zero with Gaussian distribution (de los Campos et al, 2013)

BL: Bayesian LASSO: uses an exponential prior on marker variances priors, leading to double exponential distribution of marker effects (Park & Casella 2008)

BA: Bayes A uses a scaled-t prior distribution of marker effects. (Meuwissen et al 2001).

BB: Bayes B, uses a mixture of distribution with a point mass at zero and with a slab of non-zero marker effects with a scaled-t distribution (Habier et al 2011).

BC: Bayes C same as Bayes B with a slab with Gaussian distribution.

A more detailed description of these methods can be found in Perez & de los Campos 2014 (http://genomics.cimmyt.org/BGLR-extdoc.pdf). Three semi-parametric methods:

RKHS: reproductive kernel Hilbert space and multiple kernel MRKHS, using BGLR (Gianola and van Kaam 2008). Based on genetic distance and a kernel function to regulate the distribution of marker effects. This methods is claimed to be effective for detecting non additive effects.

RF: Random forest regression, using randomForest library (Breiman, 2001, Breiman and Cutler 2013). This methods uses regression models on tree nodes which are rooted in bootstrapping data. Supposed to be able to capture interactions between markers

SVM: support vector machine, run by e1071 library. For details, see Chang, Chih-Chung and Lin, Chih-Jen: LIBSVM: a library for Support Vector Machines http://www.csie.ntu.edu.tw/~cjlin/libsvm

BRNN: Bayesian Regularization for feed-forward Neural Network, with the R-package BRNN (Gianola et al 2011). To keep computing time in reasonable limits, the parameters for the brnn function are neurons=2 and epochs = 20.

nFolds
Number of folds for the cross-validation. Smallest value recommended is nFolds = 3.

nTimes
Number of independent replicates for the cross-validation. Smallest value recommended is nTimes = 3.

Value

The class bwgs.cv returns a list containing:

summary: Summary of cross-validation, including mean and standard deviation of predictive ability (i.e. correlation between phenotype and GEBV, estimated on the validation fold, then averaged over replicates (nTimes), Time taken by the computation and number of markers

cv: Vector of predictive abilities averaged over nFolds, for each of the nTimes replicates

sd: Standard deviation of the nTimes predictive abilities

MSEP: Square root of the mean-squared error of prediction, averaged over Ntimes

SDMSEP: Standard deviation of the Square root of the mean-squared error of prediction, averaged over Ntimes

bv_table: Matrix of dimension n x 4. Columns are:

Real BV, i.e. pheno vector

Predict BV: the nx1 vector of GEBVs

gpreSD: Standart deviation of estimated GEBV

CD: coefficient of determination for each GEBV, estimated as sqrt Note that gpredSD and CD are only available for methods using the BGLR library, namely GBLUP, EGBLUP, BA,BB,BC,BL,RKHS and MKRKHS. These two columns contain NA for methods RF, RR, LASSO, EN and SVM.

Examples

# \donttest{
data(inra)
# Cross validation using GBLUP method
cv_gblup <- bwgs.cv(TRAIN47K, YieldBLUE, 
                       geno.impute.method = "mni",
                       predict.method = "gblup", 
                       nFolds = 10, 
                       nTimes = 1)

#> 2017 BWGS  - Version 1.10.0 Release date: 31/10/2017

#> Start time:

#> [1] "2020-08-03 01:39:20 UTC"

#> 

#> Warning: NAs introduced by coercion

#> Number of common lines between geno and pheno

#> [1] 100

#> Number of markers after filtering

#> [1]  100 9417

#> random pop size not applied

#> [1] 0

#> 

#> POPULATION SAMPLING METHOD

#> [1] "NULL"

#> 

#> Imputed by MNI.

#> Time of imputation by MNI (mins):

#> [1] 0.01

#> A part 5x20 of Imputed genotypic matrix:

#>         AX-89744268 AX-89622865 AX-89360616 AX-89646106 AX-89376115 AX-89396818
#> Koreli            1          -1          -1          -1          -1           1
#> CF03104           1          -1          -1          -1          -1           1
#> Barok             1          -1          -1          -1          -1          -1
#> CF99031           1          -1           1          -1          -1           1
#> EM02209          -1          -1          -1          -1          -1           1
#>         AX-89372778 AX-89318796 AX-89560313 AX-89743189 AX-89749345 AX-89378706
#> Koreli            1           1          -1          -1          -1           1
#> CF03104           0          -1          -1          -1          -1           1
#> Barok            -1          -1           1          -1           1           1
#> CF99031           1          -1           1          -1          -1          -1
#> EM02209           1           1           1          -1           1          -1
#>         AX-89349809 AX-89636031 AX-89324486 AX-89432902 AX-89654198 AX-89566186
#> Koreli            1           1          -1          -1           1          -1
#> CF03104          -1          -1          -1          -1           1           1
#> Barok            -1          -1          -1           1           1          -1
#> CF99031          -1          -1          -1           1           1          -1
#> EM02209           1          -1           1           1           1          -1
#>         AX-89563447 AX-89711500
#> Koreli            1          -1
#> CF03104          -1          -1
#> Barok            -1          -1
#> CF99031           1          -1
#> EM02209          -1           1

#> 

#> Imputed by MNI, EMI...finished.

#> No reduction for genomic data.

#> Predict by GBLUP...

#> Random Pop Size  is not applied.

#> Warning: NaNs produced

#> 

#> cv correlation for fold 1 is: 0.489

#> Warning: NaNs produced

#> 

#> cv correlation for fold 2 is: 0.402

#> Warning: NaNs produced

#> 

#> cv correlation for fold 3 is: 0.618

#> Warning: NaNs produced

#> 

#> cv correlation for fold 4 is: -0.066

#> Warning: NaNs produced

#> 

#> cv correlation for fold 5 is: 0.704

#> 

#> cv correlation for fold 6 is: 0.304

#> Warning: NaNs produced

#> 

#> cv correlation for fold 7 is: 0.303

#> Warning: NaNs produced

#> 

#> cv correlation for fold 8 is: 0.666

#> Warning: NaNs produced

#> 

#> cv correlation for fold 9 is: 0.101

#> Warning: NaNs produced

#> 

#> cv correlation for fold 10 is: 0.098

#> 

#>   |                                                |                                        |   0%  |                                                |**********                              |  25%  |                                                |*************************************   |  92%  |                                                |***********************************     |  87%  |                                                |**********                              |  25%  |                                                |****************                        |  40%  |                                                |*******************************         |  77%  |                                                |*****                                   |  12%  |                                                |********                                |  19%  |                                                |*******                                 |  16%  |                                                |***************************             |  66%  |                                                |**********************************      |  86%  |                                                |*************************************   |  93%  |                                                |**********************                  |  55%  |                                                |***********************                 |  58%  |                                                |***************************             |  69%  |                                                |**********                              |  24%  |                                                |**                                      |   4%  |                                                |************************************    |  91%  |                                                |***                                     |   7%  |                                                |****************************************| 100%  |                                                |************************                |  61%  |                                                |*******                                 |  17%  |                                                |************************************    |  91%  |                                                |*                                       |   4%  |                                                |************************                |  59%  |                                                |*********                               |  24%  |                                                |************************************    |  91%  |                                                |*********************************       |  82%  |                                                |****************************            |  70%  |                                                |*********                               |  22%  |                                                |*****************************           |  73%  |                                                |*********                               |  22%  |                                                |******************                      |  46%  |                                                |*************                           |  33%  |                                                |***********************                 |  57%  |                                                |**********                              |  25%  |                                                |*******************                     |  46%  |                                                |*************************************   |  92%  |                                                |*************************************** |  97%  |                                                |*********************************       |  82%  |                                                |************************************    |  90%  |                                                |***********************                 |  58%  |                                                |*******************************         |  77%  |                                                |****************************************| 100%  |                                                |****************************            |  71%  |                                                |*********                               |  21%  |                                                |************                            |  29%  |                                                |*****************************           |  72%  |                                                |***********************************     |  87%  |                                                |**********                              |  24%  |                                                |                                        |   0%  |                                                |**************************************  |  94%  |                                                |******************                      |  44%  |                                                |******************************          |  75%  |                                                |***************************             |  67%  |                                                |****************                        |  41%  |                                                |**************                          |  35%  |                                                |******************************          |  74%  |                                                |***************************             |  66%  |                                                |***                                     |   9%  |                                                |**********************************      |  86%  |                                                |***                                     |   8%  |                                                |**********************************      |  85%  |                                                |****                                    |  11%  |                                                |*******************                     |  48%  |                                                |**********                              |  25%  |                                                |***************************             |  69%  |                                                |*******                                 |  16%  |                                                |**************************************  |  95%  |                                                |*************                           |  32%  |                                                |**************                          |  36%  |                                                |************************************    |  89%  |                                                |*********************************       |  83%  |                                                |****                                    |  10%  |                                                |************************************    |  91%  |                                                |*******************************         |  77%  |                                                |***************                         |  38%  |                                                |****************************************| 100%  |                                                |**************                          |  35%  |                                                |**************************************  |  95%  |                                                |*********                               |  22%  |                                                |*                                       |   3%  |                                                |******                                  |  15%  |                                                |**********************************      |  85%  |                                                |*********                               |  21%  |                                                |********                                |  21%  |                                                |**                                      |   4%  |                                                |**************************************  |  94%  |                                                |**********                              |  24%  |                                                |*******************************         |  78%  |                                                |************                            |  29%  |                                                |***********************************     |  88%  |                                                |************                            |  30%  |                                                |*************************************** |  98%  |                                                |************************                |  59%  |                                                |******************************          |  76%  |                                                |*********************************       |  84%  |                                                |*******************************         |  76%  |                                                |*****************                       |  42%  |                                                |******                                  |  14%  |                                                |****************************************| 100%

#> Warning: number of items to replace is not a multiple of replacement length

#> 

#> Mean CV correlation for time 1 and 10 folds is: 0.31

#> 

#> Predict by GBLUP...ended.

#>   Mean Correlation CVs Standard Deviation CVs      Time taken (mins) 
#>                   0.31                     NA                   0.34 
#>      Number of markers 
#>                9417.00 

#> 

#> Processing...ended! Results in: object$summary, object$cv, object$sd, object$bv

# }