Computes the GEBV prediction for the target population with only genotypic Data using the options for model selection.

Computes the GEBV prediction for the target population with only genotypic Data using the options for model selection.

bwgs.predict(
  geno_train,
  pheno_train,
  geno_target,
  FIXED_train = "NULL",
  FIXED_target = "NULL",
  MAXNA = 0.2,
  MAF = 0.05,
  geno.reduct.method = "NULL",
  reduct.size = "NULL",
  r2 = "NULL",
  pval = "NULL",
  MAP = NULL,
  geno.impute.method = "NULL",
  predict.method = "GBLUP"
)

Arguments

geno_train
Matrix (n x m) of genotypes for the training population: n lines with m markers. Genotypes should be coded as -1, 0, 1, NA. Missing data are allowed and coded as NA.

pheno_train
Vector (n x 1) of phenotype for the training phenotypes. This vector should have no missing values. Otherwise, missing values (NA) will be omitted in both pheno_train and geno_train.

geno_target
Matrix (z x m) of genotypes for the target population: z lines with the same m markers as in geno_train. Genotypes should be coded as -1, 0, 1, NA. Missing data are allowed and coded as NA. Other arguments are identical to those of bwgs.cv, except pop_reduct_method, nTimes and nFolds, since the prediction is run only once, using the whole training population for model estimation, then applied to the target population.

FIXED_train
A matrix of fixed effect for training, to be used with some methods such as those included in BGLR, MUST have same rownames as geno and coded(-1 0 1)

FIXED_target
A matrix of fixed effect for targeting, to be used with some methods such as those included in BGLR, MUST have same rownames as geno and coded(-1 0 1)

MAXNA
The maximum proportion of missing value which is admitted for filtering marker columns in geno. Default value is 0.2

MAF
The minimum allele frequency for filtering marker colums in geno; default value is 0.05

geno.reduct.method
Allows sampling a subset of markers for speeding up computing time and/or avoid introducing more noise than informative markers. Options are:

RMR: Random sampling (without replacement) of a subset of markers. To be used with the parameter “reduct.marker.size”.

LD (with r2 and MAP): enables “pruning” of markers which are in LD > r2. Only the marker with the least missing values is kept for each pair in LD>r2. To allow faster computation, r2 is estimated chromosome by chromosome, so a MAP file is required with information of marker assignation to chromosomes. The MAP file should contain at least three columns: marker_name, chromosome_name and distance_from_origin (either genetic of physical distance, only used for sorting markers, LD being re-estimated from marker Data).

ANO (with pval): one-way ANOVA are carried out with R function lm on trait “pheno”. Every markers are tested one at a time, and only markers with pvalue<pval are kept for GEBV prediction.

ANO+LD (with pval and r2, MAP is facultative): combines a first step of marker selection with ANO, then a second step of pruning using LD option.

reduct.size
Specifies the number of markers for the genotypic reduction using RMR (reduct.size < m).

r2
Coefficient of linkage disequilibrium (LD). Setting 0<r2<1 if the genotypic reduction method is in LD or ANO+LD .

pval
p value for ANO method, 0 < pval < 1.

MAP
A file with markers in rows ane at least ONE columns with colnames= "chrom". Used for computing r2 within linkage groups.

geno.impute.method
Allow missing marker data imputation using the two methods proposed in function A.mat of package rrBLUP, namely:

MNI: missing data are replaced by the mean allele frequency of the marker (column in geno)

EMI: missing data are replaced using an expectation-maximization methods described in function A.mat (Endelman & Janninck 2012).

Default value is NULL.

Note that these imputation methods are only suited when there are a few missing value, typically in marker data from SNP chips of KasPAR. They are NOT suited for imputing marker data from low density to high density designs, and when there are MANY missing Data as typically provided by GBS. More sophisticated software (e.g. Beagles, Browning & Browning 2016) should be used before BWGS.

predict.method
The options for genomic breeding value prediction methods. The available options are:

GBLUP: performs G-BLUP using a marker-based relationship matrix, implemented through BGLR R-library. Equivalent to ridge regression (RR-BLUP) of marker effects.

EGBLUP: performs EG-BLUP, i.e. BLUP using a "squared" relationship matrix to model epistatic 2x2 interactions, as described by Jiang & Reif (2015), using BGLR library

RR: ridge regression, using package glmnet. In theory, strictly equivalent to gblup.

LASSO: Least Absolute Shrinkage and Selection Operator is another penalized regression methods which yield more shrinked estimates than RR. Run by glmnet library.

EN: Elastic Net (Zou and Hastie, 2005), which is a weighted combination of RR and LASSO, using glmnet library

Several Bayesian methods, using the BGLR library:

BRR: Bayesian ridge regression: same as rr-blup, but bayesian resolution. Induces homogeneous shrinkage of all markers effects towards zero with Gaussian distribution (de los Campos et al, 2013)

BL: Bayesian LASSO: uses an exponential prior on marker variances priors, leading to double exponential distribution of marker effects (Park & Casella 2008)

BA: Bayes A uses a scaled-t prior distribution of marker effects. (Meuwissen et al 2001).

BB: Bayes B, uses a mixture of distribution with a point mass at zero and with a slab of non-zero marker effects with a scaled-t distribution (Habier et al 2011).

BC: Bayes C same as Bayes B with a slab with Gaussian distribution.

A more detailed description of these methods can be found in Perez & de los Campos 2014 (http://genomics.cimmyt.org/BGLR-extdoc.pdf). Three semi-parametric methods:

RKHS: reproductive kernel Hilbert space and multiple kernel MRKHS, using BGLR (Gianola and van Kaam 2008). Based on genetic distance and a kernel function to regulate the distribution of marker effects. This methods is claimed to be effective for detecting non additive effects.

RF: Random forest regression, using randomForest library (Breiman, 2001, Breiman and Cutler 2013). This methods uses regression models on tree nodes which are rooted in bootstrapping data. Supposed to be able to capture interactions between markers

SVM: support vector machine, run by e1071 library. For details, see Chang, Chih-Chung and Lin, Chih-Jen: LIBSVM: a library for Support Vector Machines http://www.csie.ntu.edu.tw/~cjlin/libsvm

BRNN: Bayesian Regularization for feed-forward Neural Network, with the R-package BRNN (Gianola et al 2011). To keep computing time in reasonable limits, the parameters for the brnn function are neurons=2 and epochs = 20.

Value

The object bwgs.predict returns Matrix of dimension nx3. Columns are:

Predict BV: the nx1 vector of GEBVs for the validation set (rows of geno_valid)

gpredSD: Standart deviation of estimated GEBV

CD: coefficient of determination for each GEBV, estimated as sqrt ((1-stdev(GEBVi))^2/2g)

Note that gpredSD and CD are only available for methods using the BGLR library, namely GBLUP, EGBLUP, BA,BB,BC,BL,RKHS and MKRKHS. These two columns contain NA for methods RF, RR, LASSO, EN and SVM.

Examples

# \donttest{
data(inra)
# Prediction using GBLUP method
predict_gblup <- bwgs.predict(geno_train = TRAIN47K,
     pheno_train = YieldBLUE,
     geno_target = TARGET47K,
     MAXNA = 0.2, 
     MAF = 0.05,
     geno.reduct.method = "NULL",
     reduct.size = "NULL",
     r2 = "NULL",
     pval = "NULL",
     MAP = "NULL",
     geno.impute.method = "MNI",
     predict.method = "GBLUP")
#> 2017 BWGS  - Version 1.6.0 Release date: 31/02/2017
#> Start time:
#> [1] "2021-09-10 02:54:11 UTC"
#> 
#> Number of common lines between geno and pheno
#> [1] 100
#> Number of markers after filtering
#> [1]  200 8895
#> Imputed by MNI.
#> Time of imputation by MNI (mins):
#> [1] 0.01
#> A part 5x20 of Imputed genotypic matrix:
#>         AX-89744268 AX-89360616 AX-89646106 AX-89376115 AX-89396818 AX-89372778
#> Koreli            1          -1          -1          -1           1           1
#> CF03104           1          -1          -1          -1           1           0
#> Barok             1          -1          -1          -1          -1          -1
#> CF99031           1           1          -1          -1           1           1
#> EM02209          -1          -1          -1          -1           1           1
#>         AX-89318796 AX-89560313 AX-89743189 AX-89749345 AX-89378706 AX-89349809
#> Koreli            1          -1          -1          -1           1           1
#> CF03104          -1          -1          -1          -1           1          -1
#> Barok            -1           1          -1           1           1          -1
#> CF99031          -1           1          -1          -1          -1          -1
#> EM02209           1           1          -1           1          -1           1
#>         AX-89636031 AX-89673942 AX-89324486 AX-89432902 AX-89654198 AX-89566186
#> Koreli            1          -1          -1          -1           1          -1
#> CF03104          -1          -1          -1          -1           1           1
#> Barok            -1          -1          -1           1           1          -1
#> CF99031          -1          -1          -1           1           1          -1
#> EM02209          -1          -1           1           1           1          -1
#>         AX-89563447 AX-89711500
#> Koreli            1          -1
#> CF03104          -1          -1
#> Barok            -1          -1
#> CF99031           1          -1
#> EM02209          -1           1
#> 
#> Imputed by MNI, EMI...finished.
#> No reduction for genomic data.
#> Predict by GBLUP...
#> Warning: NaNs produced
#> Phenotypic estimation GEBV:
#> 
#>              gpred gpredSD    CD
#> 09BH033H041 88.223   2.395 0.386
#> 09BH025H058 89.391   2.266 0.488
#> 09BH009H003 89.786   2.667   NaN
#> 09BH032H017 91.824   2.356 0.420
#> 09BH002H021 96.315   2.482 0.294
#> 09BH039H027 88.831   2.459 0.321
#> 09BH019H047 92.625   2.707   NaN
#> 09BH007H070 88.852   2.755   NaN
#> 09BH018H014 89.271   2.718   NaN
#> 09BH032H025 92.915   2.494 0.278
#> 09BH007H073 91.950   2.788   NaN
#> 09BH039H030 89.224   2.359 0.418
#> 09BH025H059 89.844   2.217 0.520
#> 09BH038H013 90.909   2.338 0.435
#> 09BH040H062 91.269   2.565 0.154
#> 09BH033H040 89.705   2.508 0.258
#> 09BH018H009 90.647   2.554 0.180
#> 09BH002H019 92.447   2.817   NaN
#> 09BH032H021 94.021   2.262 0.491
#> 09BH019H046 90.742   2.491 0.283
#> 09BH032H019 88.294   2.332 0.440
#> 09BH019H050 89.722   2.498 0.272
#> 09BH038H009 89.546   2.465 0.314
#> 09BH039H025 90.003   2.731   NaN
#> 09BH032H024 90.706   2.395 0.387
#> 09BH025H056 87.831   2.284 0.476
#> 09BH040H066 91.948   2.335 0.437
#> 09BH018H008 91.107   2.499 0.272
#> 09BH006H048 90.133   2.779   NaN
#> 09BH006H055 92.310   2.887   NaN
#> 09BH007H069 90.374   2.668   NaN
#> 09BH032H018 89.733   2.360 0.417
#> 09BH018H007 92.505   2.653   NaN
#> 09BH002H020 92.750   2.632   NaN
#> 09BH033H039 89.559   2.384 0.396
#> 09BH038H012 91.997   2.256 0.495
#> 09BH007H072 92.164   2.724   NaN
#> 09BH006H053 91.219   2.842   NaN
#> 09BH018H011 89.587   2.653   NaN
#> 09BH025H057 89.561   2.414 0.369
#> 09BH040H058 92.300   2.397 0.385
#> 09BH033H038 88.814   2.345 0.429
#> 09BH019H043 88.209   2.603   NaN
#> 09BH009H002 91.766   2.668   NaN
#> 09BH006H052 95.402   2.651   NaN
#> 09BH040H060 89.853   2.190 0.537
#> 09BH019H045 88.600   2.577 0.122
#> 09BH040H065 91.731   2.183 0.541
#> 09BH040H059 90.913   2.358 0.418
#> 09BH018H010 90.004   2.712   NaN
#> 09BH002H023 94.130   2.567 0.151
#> 09BH032H022 91.449   2.380 0.400
#> 09BH038H010 89.591   2.420 0.363
#> 09BH002H017 88.974   2.625   NaN
#> 09BH025H055 89.613   2.418 0.364
#> 09BH019H044 90.394   2.274 0.483
#> 09BH002H018 91.743   2.685   NaN
#> 09BH032H026 89.414   2.310 0.457
#> 09BH018H006 90.060   2.503 0.266
#> 09BH040H063 90.038   2.453 0.328
#> 09BH006H050 92.966   2.873   NaN
#> 09BH033H044 88.896   2.451 0.330
#> 09BH038H005 89.651   2.319 0.450
#> 09BH007H074 90.367   2.762   NaN
#> 09BH039H033 89.637   2.570 0.143
#> 09BH006H049 90.512   2.656   NaN
#> 09BH019H048 94.114   2.392 0.389
#> 09BH040H064 92.358   1.743 0.741
#> 09BH039H026 91.515   2.314 0.453
#> 09BH033H045 90.159   2.608   NaN
#> 09BH039H028 90.275   2.299 0.465
#> 09BH002H015 95.189   2.721   NaN
#> 09BH006H051 88.054   2.765   NaN
#> 09BH039H032 91.171   2.483 0.292
#> 09BH033H046 91.047   2.528 0.228
#> 09BH019H053 88.916   2.545 0.198
#> 09BH038H007 91.841   2.310 0.456
#> 09BH025H051 90.217   2.413 0.369
#> 09BH018H013 92.444   2.644   NaN
#> 09BH033H042 90.019   2.575 0.129
#> 09BH038H006 90.306   2.506 0.262
#> 09BH033H043 88.717   2.508 0.258
#> 09BH025H054 89.868   2.185 0.540
#> 09BH002H025 92.584   2.832   NaN
#> 09BH039H029 91.264   2.316 0.452
#> 09BH040H061 91.008   2.435 0.347
#> 09BH038H011 87.746   2.234 0.510
#> 09BH038H008 90.874   2.305 0.460
#> 09BH025H053 89.614   2.415 0.367
#> 09BH002H016 94.830   2.708   NaN
#> 09BH002H014 93.115   2.737   NaN
#> 09BH018H012 90.792   2.555 0.178
#> 09BH032H023 93.040   2.345 0.430
#> 09BH025H050 89.938   2.331 0.441
#> 09BH007H071 91.431   2.695   NaN
#> 09BH009H001 92.399   2.473 0.304
#> 09BH006H047 91.863   2.946   NaN
#> 09BH039H031 91.773   2.037 0.620
#> 09BH019H049 89.672   2.604   NaN
#> 09BH006H054 92.821   2.570 0.143
#> 
#> Stop time:
#> [1] "2021-09-10 02:54:13 UTC"
#> Time difference of 2.198257 secs
# }